+++ src/Chunkyard/Core/AesGcmCrypto.cs
@@ -6,6 +6,7 @@ namespace Chunkyard.Core;
 public class AesGcmCrypto
 {
     public const int DefaultIterations = 1000;
+    public const int EncryptionBytes = NonceBytes + TagBytes;
 
     private const int NonceBytes = 12;
     private const int TagBytes = 16;
@@ -26,6 +27,32 @@ public class AesGcmCrypto
 
     public int Iterations { get; }
 
+    public int Encrypt(
+        ReadOnlySpan<byte> nonce,
+        ReadOnlySpan<byte> plainText,
+        byte[] buffer)
+    {
+        ArgumentNullException.ThrowIfNull(buffer);
+
+        nonce.CopyTo(
+            new Span<byte>(buffer, 0, nonce.Length));
+
+        var ciphertext = new Span<byte>(
+            buffer,
+            nonce.Length,
+            plainText.Length);
+
+        var tag = new Span<byte>(
+            buffer,
+            nonce.Length + ciphertext.Length,
+            TagBytes);
+
+        using var aesGcm = new AesGcm(_key);
+        aesGcm.Encrypt(nonce, plainText, ciphertext, tag);
+
+        return nonce.Length + plainText.Length + TagBytes;
+    }
+
     public byte[] Encrypt(
         ReadOnlySpan<byte> nonce,
         ReadOnlySpan<byte> plainText)
@@ -33,21 +60,7 @@ public class AesGcmCrypto
         var cipherText = new byte[
             nonce.Length + plainText.Length + TagBytes];
 
-        nonce.CopyTo(
-            new Span<byte>(cipherText, 0, nonce.Length));
-
-        var innerCiphertext = new Span<byte>(
-            cipherText,
-            nonce.Length,
-            plainText.Length);
-
-        var tag = new Span<byte>(
-            cipherText,
-            cipherText.Length - TagBytes,
-            TagBytes);
-
-        using var aesGcm = new AesGcm(_key);
-        aesGcm.Encrypt(nonce, plainText, innerCiphertext, tag);
+        Encrypt(nonce, plainText, cipherText);
 
         return cipherText;
     }