+++ src/Chunkyard/GlobalUsing.cs
@@ -10,6 +10,8 @@ global using System.Security.Cryptography;
 global using System.Text;
 global using System.Text.Json;
 global using System.Text.RegularExpressions;
+global using System.Threading;
+global using System.Threading.Tasks;
 global using Chunkyard.Core;
 global using Chunkyard.Infrastructure;
 global using CommandLine;